DANN/core/autogen.py
"""AutoGen connector scaffold.

This module provides a minimal wrapper for AutoGen-style multi-agent runs.
If `autogen` package is not installed, operations will return stubs.
"""
import os
import logging

try:
    import openai  # placeholder for whatever autogen requires
    HAS_AUTOGEN = True
except Exception:
    HAS_AUTOGEN = False

logger = logging.getLogger(__name__)


class AutoGenClient:
    def __init__(self, api_key: str = None):
        self.api_key = api_key or os.environ.get('AUTOGEN_API_KEY')
        if not HAS_AUTOGEN:
            logger.warning('AutoGen support not installed. AutoGenClient will run in stub mode.')

    def run_agents(self, agents_spec: dict, input_data: dict):
        """Run a multi-agent AutoGen-style workflow.

        This is a scaffold returning a stub when real AutoGen is not installed.
        """
        if not HAS_AUTOGEN:
            logger.info('Stub: would run AutoGen with agents_spec=%s and input=%s', agents_spec, input_data)
            return {"status": "stub", "agents": list(agents_spec.keys()), "input": input_data}

        try:
            # Pseudo-code for real integration
            # from autogen import MultiAgentSystem
            # mas = MultiAgentSystem(agents_spec)
            # out = mas.run(input_data)
            out = {"status": "ok", "result": {}}
            return out
        except Exception as e:
            logger.exception('AutoGen run failed: %s', e)
            return {"status": "error", "error": str(e)}
